feat: 주문 목록 리스트 아이템 로직별 컴포넌트 분리 및 공통 타입 작성
